% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{correct.overflow.filter}
\alias{correct.overflow.filter}
\title{Correct the overflow in filter names}
\usage{
correct.overflow.filter(df)
}
\arguments{
\item{df}{A data frame containing overflow filter names.}
}
\value{
A data frame with overflow from service names corrected to be filter names
}
\description{
Correct the overflow in filter names
}
\examples{
API.tables <- get.all.tables()
df <- API.tables[[5]] 
corrected.df <- correct.overflow.filter( df )
corrected.df
}
