% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{create.authentication}
\alias{create.authentication}
\title{Combine email and key into one string for querying}
\usage{
create.authentication()
}
\description{
Combine email and key into one string for querying
}
\examples{
create.authentication()
AUTHENTICATION
}
