% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{create.filter.list}
\alias{create.filter.list}
\title{Create the filter list to allow for chained variable calls}
\usage{
create.filter.list(df)
}
\arguments{
\item{df}{Data frame containing the service. Assumes filters exist for this data frame.}
}
\value{
A list for filters in a service
}
\description{
Create the filter list to allow for chained variable calls
}
\examples{
API.tables <- get.all.tables()
df <- API.tables[[4]]
filter.list <- create.filter.list( df )
filter.list
}
