% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{create.list.endpoints.examples.params}
\alias{create.list.endpoints.examples.params}
\title{Make a list containing endpoints, example, and parameters for a service}
\usage{
create.list.endpoints.examples.params(row.number, df)
}
\arguments{
\item{row.number}{The row number in the data frame containing a service}

\item{df}{A data frame containing services}
}
\value{
A list structure to allow for chained variable calls
}
\description{
Make a list containing endpoints, example, and parameters for a service
}
\examples{
API.tables <- get.all.tables()
df <- API.tables[[4]]
row.number <- 2
result <- create.list.endpoints.examples.params( row.number, df)
result
}
