% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{determine.filter}
\alias{determine.filter}
\title{Find the indices for a filter in a data frame}
\usage{
determine.filter(name.filter, df)
}
\arguments{
\item{name.filter}{The name of the filter we want}

\item{df}{Data frame containing the filters}
}
\value{
Vector of integer indices showing where a particular filter is
}
\description{
Find the indices for a filter in a data frame
}
\examples{
API.tables <- get.all.tables()
df <- API.tables[[4]]
name.filter <- 'Is the API available for use?'
indices <- determine.filter( name.filter, df)
indices
}
