% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{example.check}
\alias{example.check}
\title{Find out if a string has 'Example' in it}
\usage{
example.check(string)
}
\arguments{
\item{string}{A string, intended as an entry in a dataframe containing info about services}
}
\value{
Boolean reflecting presence of 'Example' in string
}
\description{
Find out if a string has 'Example' in it
}
\examples{
example.check( "Example number one")
example.check( "Number two ")
}
