% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{get.list.variable.endpoint}
\alias{get.list.variable.endpoint}
\title{Show endpoint for listing information on a variable}
\usage{
get.list.variable.endpoint(variable.type)
}
\arguments{
\item{variable.type}{A variable used in the EPA API service. Consult VARIABLE.TYPES for available variables.}
}
\value{
An endpoint character that lists information to help the user query a variable.
}
\description{
Show endpoint for listing information on a variable
}
\examples{
get.list.variable.endpoint( "state" )
get.list.variable.endpoint( "classes" )
}
