% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{get.service.names}
\alias{get.service.names}
\title{Get service names and descriptions to the services}
\usage{
get.service.names()
}
\value{
A data frame containing services with names and descriptions offered by the EPA API.
}
\description{
Get service names and descriptions to the services
}
\examples{
get.service.names()
SERVICE.NAMES
}
