% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{get.variables}
\alias{get.variables}
\title{Popoulate VARIABLES for info on making requests}
\usage{
get.variables()
}
\value{
Data frame containing variables and information about them used in the EPA API.
}
\description{
Popoulate VARIABLES for info on making requests
}
\examples{
vars <- get.variables()
vars$edate
}
