% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{list.remove.escapes.spaces}
\alias{list.remove.escapes.spaces}
\title{Remove '\t', '\r\n', "   " from entries in a list}
\usage{
list.remove.escapes.spaces(a.list)
}
\arguments{
\item{a.list}{List to remove entries from.}
}
\value{
A list without '\t', '\r\n' and "  "
}
\description{
Remove '\t', '\r\n', "   " from entries in a list
}
\examples{
get.services()
SERVICES <- list.remove.escapes.spaces( SERVICES )
SERVICES
}
