% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{perform.call.raw}
\alias{perform.call.raw}
\title{Perform call and maintain jsonlite structure}
\usage{
perform.call.raw(call)
}
\arguments{
\item{call}{URL following structure from EPA API}
}
\value{
Results of data request in json format
}
\description{
Perform call and maintain jsonlite structure
}
\examples{
endpoint <- 'list/states'
call <- create.base.call( endpoint )
raw.call <- perform.call.raw( call )
raw.call
}
