% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{perform.call}
\alias{perform.call}
\title{Perform call and convert data into list}
\usage{
perform.call(call)
}
\arguments{
\item{call}{URL following structure from EPA API}
}
\value{
A list containing requested data
}
\description{
Perform call and convert data into list
}
\examples{
endpoint <- 'list/states'
call <- create.base.call( endpoint )
call.data <- perform.call( call )
call.data$Header
call.data$Data
}
