% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{populate.all.services}
\alias{populate.all.services}
\title{Take a list of html tables, output a list of lists, each list a service}
\usage{
populate.all.services(list.tables)
}
\arguments{
\item{list.tables}{EPA API html tables}
}
\value{
A list containing services, each service is a list.
}
\description{
Take a list of html tables, output a list of lists, each list a service
}
\examples{
API.tables <- get.all.tables()
services <- populate.all.services( API.tables )
services$MetaData$Filter$`Known Issues`$Endpoint
}
