% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{populate.service.list}
\alias{populate.service.list}
\title{Take a df - a table from EPA site - and output a list with easy accessing for variables}
\usage{
populate.service.list(df)
}
\arguments{
\item{df}{Data frame of an HTML table in the EPA API site.}
}
\value{
A list containing the service variables for a service in the EPA API site
}
\description{
Take a df - a table from EPA site - and output a list with easy accessing for variables
}
\examples{
API.tables <- get.all.tables()
df <- API.tables[[4]]
result <- populate.service.list( df )
result$Filter$`Known Issues`$RequiredVariables
}
