% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{remove.all.service.names}
\alias{remove.all.service.names}
\title{Remove extraneous service name variable from all services}
\usage{
remove.all.service.names(services)
}
\arguments{
\item{services}{The EPA API services as a list}
}
\value{
List of services without service name variable
}
\description{
Remove extraneous service name variable from all services
}
\examples{
SERVICES <- remove.all.service.names( SERVICES )
SERVICES
}
