% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{remove.service.name}
\alias{remove.service.name}
\title{Remove extraneaous service name variable in a service list}
\usage{
remove.service.name(service)
}
\arguments{
\item{services}{List of services from the EPA API}
}
\value{
A service list without the service name variable
}
\description{
Remove extraneaous service name variable in a service list
}
\examples{
services <- SERVICES
services[[1]] <- remove.service.name( services[[1]] )
services[[1]]

}
