% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{service.names}
\alias{service.names}
\title{Names of services offered by the EPA API}
\description{
The service.names list contains names of all services offered by the EPA API
along with a description of each service.
}
\keyword{datasets}
