% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{services}
\alias{services}
\title{Services offered by the EPA API}
\description{
The services list contains comprehensive information about all services provided
by the EPA API site.
}
\keyword{datasets}
