% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{set.email}
\alias{set.email}
\title{Email for API in queries}
\usage{
set.email(email)
}
\arguments{
\item{email}{Email used to sign up with the EPA API. 
Sign your email up at https://aqs.epa.gov/aqsweb/documents/data_api.html#signup.}
}
\description{
Email for API in queries
}
\examples{
set.email( "myemail@domain.com" )
EMAIL 
}
