% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{set.key}
\alias{set.key}
\title{Key for API in queries}
\usage{
set.key(key)
}
\arguments{
\item{key}{Key for making data request to EPA API.
Get your key at https://aqs.epa.gov/aqsweb/documents/data_api.html#signup.}
}
\description{
Key for API in queries
}
\examples{
set.key( "mykey" )
KEY
}
