% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAGrabberUtility.R
\name{setup.services.variables.endpoints}
\alias{setup.services.variables.endpoints}
\title{Populate SERVICES, VARIABLES, and ENDPOINTS to be ready for the user to query}
\usage{
setup.services.variables.endpoints()
}
\description{
Populate SERVICES, VARIABLES, and ENDPOINTS to be ready for the user to query
}
\examples{
setup.services.variables.endpoints()
SERVICES
VARIABLE
ENDPOINTS
}
